feat: The start.bat file was added to improve project initialization and syntax and styling corrections were made
